requirements.txt
Django==1.11.3
psycopg2==2.7.1
Django==1.11.3
djangorestframework==3.6.2
djangorestframework-jsonapi==2.2.0
django-safedelete==0.4.5
django-adminplus==0.5
Markdown==2.6.8
pyjwt==1.5.2
